Get rid of hacked Wicket/Rome versions
Let's use standard distributions instead.

From looking at the Wicket code and the original hack at
https://github.com/gerritcodereview/wicket/commit/4a3f2c1dccf064008f223bb554c90e12c1c3b55d
it appears to me that this hack may not have been necessary in the first
place and just overriding getClassLoader() in the GerritWicketFilter
might be good enough.

Adjust both Maven and Buck driven builds.

Change-Id: I7e8a6e5a0a0decefac8e971a627111a9e827dfe7