Use a local build of GitBlit 1.7.1 for static res
Gerrit plugins need to access static resources under the /static
folder of the jar, which unfortunately is different from what is
typically done when distributing GitBlit jar on Maven.

We need to build GitBlit 1.7.1 from sources with the /static resources
to get the GitBlit plugin working in Gerrit and serving all CSS and images.

Change-Id: I28101db85bfa40f57fc3c2dc75aff88cdd996ed9