Update patch set 5
Patch Set 5:

Happy to merge it as "code-compatibility" fix. However, it wouldn't work because of Lucene incompatibilities. That is kind of expected, because it took a lot of effort to upgrade the Gitblit ones to make them compatible with Gerrit master again.

Patch-set: 5
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>