Adapt code to recent changes on master and bump Jgit to snapshot
Use new Bazel feature: patch-in-import and build Gitblit in place,
instead of relying on the patched GitBlit to be installed in local
maven repository. The provded two patches do the following:

1. Bump JGit to version used in Gerrit master

This is needed for two reason:
* Deprecated Repository.getRef(String) method is removed
* Deprecated class org.eclipse.jgit.util.io.SafeBufferedOutputStream
is removed

2. Extend build.xml to offer buildGitblit target to create gitblit.jar

This is needed, because currently gitblit.jar is only created as a part
of installMaven target that also installs the artifact into local Maven
repository.

Change-Id: I61aa48664dfb806b46749d1e376294782ba7a6fc