Update patch set 2
Patch Set 2:

> To rectify, we would need to bump JGit in GitBlit and adapt the code to new JGit version to be compatible with JGit version used in Gerrit core.

The problem is that the same version of gitblit is used in all branches of the gitblit plugin.

It would be better if there was a stable release of gitblit that we can use, instead of building it from source ourselves.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>