Bazel: Build gitblit from source
Gitblit has many dependencies that also used in gerrit core. It turns
out, that some dependencies were used in different version, that was
not noticed. So that it was more or less luck that it worked. The only
confidental way to make sure, that all libraries are used in the same
versions is to compile gitblit against dependencies used in gerrit.
Trying to do it that way reveals many issues in the current approach:

These transitive dependencies were missing in the plugin build:

* jetty-webapp
* jcalendar
* jsoup
* force-wsc
* jna-platform
* intellij-annotations
* jedis
* commons-pool2

These libraries were used in different versions:

* args4j (update in gerrit core)
* commons-io (update in gerrit core)
* commons-compress
* Jetty
* JGit
* SSH Daemon

This change is based on forked gitblit version with dependencies update
to match the library versions used in gerrit core. Alternative approach
would be to include the patches in the gitblit-plugin repository and
apply them during the fetch procedure.

Change-Id: I47a1eae1f686e8bdac96ec488f9bbc9d72277a73